+++ CHANGELOG.md
@@ -3,4 +3,5 @@
 ## [1.0.0] - 2025-08-19
 ### Create a project
  - Added AutoIndentPaste, a basic feature.
-
+ - Supports multi-cursor paste (spread/full mode)
+ - Option to override default Ctrl/Cmd+V